Code refactoring
Abstraction was added in place of online/offline learning procedures.

Code was refactored and placed in multiple
 go files.
